Flytter vedtakdto for barnetilsyn til egen fil
